cccd_orc_api/requirements.txt
flask
uvicorn
flask_sqlalchemy
flask_jwt_extended
pillow
opencv-python==4.6.0.66
ultralytics
pytesseract
python-dotenv
pymysql
paddlepaddle==2.6.2
paddleocr==2.7.3
pytest
numpy==1.26.4
vietocr
flask_cors